Change conductivity app name to match salinity
